feat: simplify header generation for client metadata
