+++ packages/mcp/src/lib/encryption.ts
@@ -31,17 +31,44 @@ function encryptClientIp(clientIp: string): string {
   }
 }
 
-export function generateHeaders(
-  clientIp?: string,
-  apiKey?: string,
-  extraHeaders: Record<string, string> = {}
-): Record<string, string> {
-  const headers: Record<string, string> = { ...extraHeaders };
-  if (clientIp) {
-    headers["mcp-client-ip"] = encryptClientIp(clientIp);
+export interface HeaderContext {
+  clientIp?: string;
+  apiKey?: string;
+  clientInfo?: {
+    ide?: string;
+    version?: string;
+  };
+  serverVersion?: string;
+  transport?: "stdio" | "http";
+}
+
+/**
+ * Generate headers for Context7 API requests.
+ * Handles client IP encryption, authentication, and telemetry headers.
+ */
+export function generateHeaders(context: HeaderContext): Record<string, string> {
+  const headers: Record<string, string> = {
+    "X-Context7-Source": "mcp-server",
+  };
+
+  if (context.clientIp) {
+    headers["mcp-client-ip"] = encryptClientIp(context.clientIp);
   }
-  if (apiKey) {
-    headers["Authorization"] = `Bearer ${apiKey}`;
+  if (context.apiKey) {
+    headers["Authorization"] = `Bearer ${context.apiKey}`;
   }
+  if (context.serverVersion) {
+    headers["X-Context7-Server-Version"] = context.serverVersion;
+  }
+  if (context.clientInfo?.ide) {
+    headers["X-Context7-Client-IDE"] = context.clientInfo.ide;
+  }
+  if (context.clientInfo?.version) {
+    headers["X-Context7-Client-Version"] = context.clientInfo.version;
+  }
+  if (context.transport) {
+    headers["X-Context7-Transport"] = context.transport;
+  }
+
   return headers;
 }
